+++ archiver/compressor/interface.go
@@ -1,4 +1,4 @@
-package format
+package compressor
 
 import (
 	"io"
