+++ main.go
@@ -2,9 +2,15 @@ package main
 
 import (
 	"flag"
+	"fmt"
+	"github.com/shinofara/stand/archiver"
 	"github.com/shinofara/stand/backup"
-	"github.com/shinofara/stand/compressor"
 	"github.com/shinofara/stand/config"
+	"time"
+)
+
+const (
+	TIME_FORMAT = "20060102150405"
 )
 
 type Args struct {
@@ -30,7 +36,9 @@ func main() {
 
 		switch cfg.Type {
 		case "dir":
-			uploadFileName, err = compressor.Compress(cfg)
+			output := makeCompressedFileName(cfg)
+			a := archiver.New(cfg)
+			uploadFileName, err = a.Archive(output)
 			if err != nil {
 				panic(err)
 			}
@@ -45,4 +53,23 @@ func main() {
 			panic(err)
 		}
 	}
+}
+
+//input, output
+func makeCompressedFileName(cfg *config.Config) string {
+	timestamp := time.Now().Format(TIME_FORMAT)
+
+	extention := "zip"
+	switch cfg.CompressionConfig.Format {
+	case "tar":
+		extention = "tar.gz"
+	}
+
+	var output string
+	if cfg.CompressionConfig.Prefix != "" {
+		output = fmt.Sprintf("%s%s.%s", cfg.CompressionConfig.Prefix, timestamp, extention)
+	} else {
+		output = fmt.Sprintf("%s.%s", timestamp, extention)
+	}
+	return "/tmp/" + output
 }
